Suppress Phan check for plain text output in UserMailer::send
UserMailer::send defaults to plain/text and only outputs HTML if
the body element is an array with an 'html' key. This extension's
email messages are always strings, which UserMailer::send
treats as plain/text.

The Phan build checks that potential HTML entities are escaped before
output, but the text output by this extension is always plain text
which does not need escaping. This leads to build failures that
need to be suppressed.

Change-Id: Id4979bc69229816e52632d24c39ec6faa798ce0f